Wait for dev_path and twin before returning
I cannot trigger this in tests but in application code the process is
started before the files exist which leads to errors when trying to
write to the files - on my 14" m2 macbook pro it takes about 2ms. This
change polls the file system for the two file's existance before
returning.