Halt `TTY0TTY.open/2` until null device is created
Fixes #1
Inspired by #3

The time to open null devices varies across systems. When using this in
unit tests, that means flaky tests between dev machine and CI.

This adds a little blocking to the `TTY0TTY.open/1` to wait until the
devices are created. This can be controlled with `timeout: ms` option
to adjust the default 5000ms